♻️ Refactor to router & routes
