+++ src/lib/routes.js
@@ -1,0 +1,7 @@
+import { renderIndex } from '../pages'
+
+const routes = {
+  '/': renderIndex,
+}
+
+export { routes }
