+++ src/main.js
@@ -1,32 +1,4 @@
-import { renderIndex } from './pages'
+import { start } from './lib/router'
+import { routes } from './lib/routes'
 
-let routes
-window.addEventListener('popstate', (event) => {
-  if (routes[location.pathname]) {
-    routes[location.pathname]()
-    return
-  }
-})
-
-const goto = (url, { push } = {}) => {
-  const pathname = url
-
-  if (push) {
-    history.pushState({}, '', url)
-  }
-  routes[pathname]()
-  return
-}
-
-const start = (params) => {
-  routes = params
-  goto(location.pathname + location.search)
-}
-
-function main() {
-  start({
-    '/': renderIndex,
-  })
-}
-
-main()
+start(routes)
